requirements.txt
Django==5.0.1
mysqlclient==2.2.1
python-dotenv==1.0.0
django-recaptcha==4.0.0
django-widget-tweaks==1.5.0
django-secured-fields==0.4.3
django-stdimage==6.0.2
Pillow==10.2.0
django-environ==0.11.2
boto3==1.28.62